ci: improvement (#3293)
* refactor: combine flake check and matrix generation

This commit combines the flake-checks and generate-matrix jobs
into a single "prepare" job to improve workflow efficiency.

* chore: update nix-installer-action and build reporting

Updates the nix-installer-action and improves build reporting.

* fix(ci): use correct attribute name for nix installer

The attribute `nix-conf` is deprecated, `extra-conf` should be used.

* fix(build): handle missing build info gracefully

If `nix path-info` fails, the workflow now continues
instead of erroring. It also sets closure_size=N/A.

* chore: simplify build workflow

Uses flake-checker-action and removes redundant build report generation.

* chore: run nix flake check in CI

This commit adds a new job to the CI workflow that runs `nix flake
check`. This will help to ensure that the flake is valid and that
all dependencies are available.